Move is_activated check to the authorization server logic.
