+++ toolbox_generics/urls.py
@@ -2,7 +2,7 @@ from django.contrib import admin
 from django.urls import include
 from django.urls import path
 
-from oauth2_provider.views import TokenView
+from .views import TokenView
 
 
 urlpatterns = [
